safari-600.7-branch/LayoutTests/editing/pasteboard/avoid-copying-body-with-background.html

<!DOCTYPE html>
<html>
<head>
<style>
body {
    background: transparent;
    width: 500px;
    height: 200px;
    overflow-x: auto; overflow-y:auto;
    text-decoration: underline;
}
</style>
</head>
<body style="color: rgb(51, 51, 51); " contenteditable="true">
<ul>
<li>hello, world <a href="http://www.webkit.org/">WebKit</a></li>
<li>Copied content</li>
</ul>
<script src="../../resources/dump-as-markup.js"></script>
<script>

document.body.focus();
document.execCommand('SelectAll', false, null);
document.execCommand('Copy', false, null);
getSelection().modify('Move', 'Forward', 'Character');
document.execCommand('Paste', false, null);

while (script = document.querySelector('script'))
    script.parentNode.removeChild(script);

Markup.description('This test copies and pastes content inside a body with background and text-decoration properties. WebKit should not nest body.');
Markup.dump(document.body);

</script>
</body>
</html>
