LayoutTests/editing/pasteboard/avoid-copying-body-with-background-expected.txt

This test copies and pastes content inside a body with background and text-decoration properties. WebKit should not nest body.
| "\n"
| <ul>
|   "\n"
|   <li>
|     "hello, world "
|     <a>
|       href="http://www.webkit.org/"
|       "WebKit"
|   "\n"
|   <li>
|     "Copied content"
|   <li>
|     "hello, world "
|     <a>
|       href="http://www.webkit.org/"
|       "WebKit"
|   <li>
|     "Copied content<#selection-caret>"
|   "\n"
| "\n"
| "\n"
